Specify multiple buildpacks now needed for deploy
Need to specify both Ruby (previously just this was auto-detected) and
Node buildpacks now after updating JS dependencies, to avoid this error
on deploy:

```
error @rails/webpacker@5.2.1: The engine "node" is incompatible with this module. Expected version ">=10.17.0". Got "10.15.3"
```

According to
http://dokku.viewdocs.io/dokku/deployment/methods/buildpacks/:

```
Always remember to pin your buildpack versions when using the
multi-buildpacks method, or you may find deploys changing your deployed
environment.
```

I'm not totally sure what this means or whether this achieves this, but
could be worth considering if this setup ever causes issues in future.