Fix issue where non-Tasks could be added to List Tasks
Since `self.tasks` was being conditionally assigned, this conversion
would not occur for an Array method that modified the Array in place but
then returned another value.

A better approach is to always ensure `self.tasks` is converted to an
Array of Tasks (I can't see an issue with doing this unnecessarily), and
then either return the new `self.tasks` or the original result,
depending on the delegated Array method's return value (so that the
result of this method as close as possible matches the delegated method,
but with any conversion still present in this return value).