Add support for `parse_code_blocks` Config option
Along with initial setup and refactoring to introduce Config, since this
is the first Config option.

When this option is set, anything within backticks in a Task will just
be parsed as text. This is not strictly compliant with the todotxt
specification but is useful to be able to include things which would
otherwise be parsed as something else within a Task (such as projects,
contexts, or colons), without this having unintended effects (e.g. see
https://github.com/bobwhitelock/todotxt-ui/issues/50).