Add support for `task_class` Config option
When passed, this allows the class used to create Tasks internally to be
customized. This will be useful to allow custom, app-specific behaviour
to be added by default to Tasks, without having to monkey-patch or jump
through hoops decorating Tasks in order to do this.

The class used should probably be a subclass of `Todotxt::Task` for this
to work well and be useful, but this isn't absolutely necessary and so
isn't enforced.