Fix bug where reloaded List would report as dirty
This was due to `reload` not resetting `original_tasks`, and so the old
value for `original_tasks` would remain after a `reload`.

Rather than just updating `reload` to also set `original_tasks`, instead
fix this by delegating to `initialize`, so this and any future behaviour
added to `initialize` will also take place on a `reload`.