Use lock file to prevent Rake tasks conflicting (maybe fixes #128)
Each task will now obtain an exclusive lock file at the start - for
`progress_scheduled_tasks` it will block for five minutes waiting on
this (the timeout should never actually be reached), while `sync_deltas`
will not block as will just run again in a minute anyway.

Also improve logging in these tasks a bit at the same time, to help with
debugging any issues with this.

Hopefully this fixes
https://github.com/bobwhitelock/todotxt-ui/issues/128, and possibly this
will be a better fix for
https://github.com/bobwhitelock/todotxt-ui/issues/112 as well - the
retry logic is probably still worthwhile though, to make this task more
robust generally.

Still not sure this is an ideal way to do all this, and this is
definitely lacking in test coverage - can consider a better approach to
robustly running these tasks in
https://github.com/bobwhitelock/todotxt-ui/issues/100, and/or test
coverage for this in
https://github.com/bobwhitelock/todotxt-ui/issues/13.