Revert "Use `spec_helper` in all parser tests"
This reverts commit f415a94300b5d72ffe1615ab09775e336e06848f.

This commit was actually preventing these tests from being able to run
when run as individual test files, as they are dependent on Rails
autoloading. I didn't notice this until now as they still pass fine when
the suite is run as a whole. Can find an alternative way of handling
these that isn't dependent on Rails if/when extract the parser to its
own repo, but for now just revert this commit.