Define full routing for multiple todo files
Have a path for each todo file (corresponding to each tab), independent
paths for add and edit under this for each todo file, and redirect `/`
and `/add` to the corresponding routes for the first todo file (which
should be the main one).

As part of this needed to rework how tabs work a lot, and seemed
simplest to collapse `Root` and `Tabs` into single new `TodoFile` page
component (this is clearer and simpler, and no need for the previous
over-abstracted `Tabs` component).