Regenerate `Gemfile.lock` to fix install error
Due to this error on `bundle install`:

```
Your bundle is locked to i18n (1.9.0) from rubygems repository https://rubygems.org/ or installed locally, but that version can no longer be found in that source. That means the author of i18n (1.9.0) has removed it. You'll need to update your bundle to a version other than
i18n (1.9.0) that hasn't been removed in order to install.
```

As suggested at
https://talk.jekyllrb.com/t/your-bundle-is-locked-to-i18n-1-9-0-from-rubygems/6950/4.